reorganized some folders, fixed homepage button color
