+++ app/index.js
@@ -3,10 +3,10 @@ import ReactDOM from 'react-dom';
 import { Router, Route, browserHistory } from 'react-router';
 import { Provider } from 'react-redux';
 import Home from './components/Home';
-import NewGoal from './components/NewGoal';
+import NewGoal from './components/newgoal/NewGoal';
 import Signup from './components/Signup';
-import PostWorkout from './components/PostWorkout';
-import Dashboard from './components/Dashboard/Dashboard';
+import PostWorkout from './components/postworkout/PostWorkout';
+import Dashboard from './components/dashboard/Dashboard';
 
 import './scss/primary.scss';
